We're coming back from the dead.
It's been several eternities, but we finally have opened the door to the abandoned laboratory. Many infinities ago, the great dan-simon was to create the fabled prestige layer of Intergalactic, but he was quickly never seen again after a tragic event. He sealed the door to the public and a trustworthy person was granted this ability to unseal it, in hopes that one day, the dream may come true. Upon unsealing the door, we are going to restore what was lost in this time of desperation; a time of fallen news tickers and in the midst of new chaos emerging from the mobile app...

This is Eternity Respecced.

INTERGALACTIC.