PROGRESS REPORT 1 - 3/10/2020
** This is the more updated lore description. **

It's been several eternities, but we finally have opened the door to the abandoned laboratory. Many infinities ago, the great dan-simon was to create the fabled prestige layer of Intergalactic, but he was quickly never seen again after a tragic event. He sealed the door to the public and a trustworthy person was granted this ability to unseal it, in hopes that one day, the dream may come true. Upon unsealing the door, we are going to restore what was lost in this time of desperation; a time of fallen news tickers and in the midst of new chaos emerging from the mobile app. A time when peace was abundant within the discord is no more. There is only despair.

We knew that this day may come... 

... and there it is. The chosen one.

usavictor unseals the door. With his newfound power of JavaScript, he is no longer afraid of what his past predecessor could only achieve. The universe he once knew has shattered, and traces of Aarex have been left everywhere. He had created the almighty document of NG+3.1, however, it had not been fully executed. It had been sent to development hell itself. usavictor needed to return to create his own sacred modules with balance.

Within the old laboratory lies a series of blueprints, each blueprint showing the new capabilities of the galactic studies and the galactic dimensions themselves. Their power was never known to anyone but the creator. He begins to understand them, surely through several perspectives. However, when he tried to test them, it didn't work. He thought, that somehow he could fix these to help restore the old projects. Then it became clear.

He knows what he must do. He must restore the balance of the universe. The universe calls him, to rebalance the world and bring peace among matter and antimatter, with the new power of galactic dimensions and studies. Well, we really don't know about matter, or 5 hours. Those mean nothing to him anymore. What matters now is how we fix it.

He begins to run the antimatter machine... and a new journey begins.

This is Eternity Respecced.

INTERGALACTIC.